Move util classes in auth-api-nongms to auth-api
Refactoring for future development on supporting more authentication provider that also uses web-based OAuth2.

Fixes #111.
