+++ auth-api/src/main/java/com/omh/android/auth/api/utils/lifecycle/LifecycleRunnable.kt
@@ -14,15 +14,12 @@
  * limitations under the License.
  */
 
-package com.omh.android.auth.nongms.utils.lifecycle
+package com.omh.android.auth.api.utils.lifecycle
 
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.LifecycleOwner
 
-internal class OnLifecycleStateObserver(private val runnable: LifecycleRunnable) : LifecycleEventObserver {
-
-    override fun onStateChanged(source: LifecycleOwner, event: Lifecycle.Event) {
-        runnable.run(source, event, this)
-    }
+fun interface LifecycleRunnable {
+    fun run(source: LifecycleOwner, event: Lifecycle.Event, observer: LifecycleEventObserver)
 }
